+++ lib/Inprint/Catalog/Readiness.pm
@@ -1,14 +1,40 @@
 package Inprint::Catalog::Readiness;
 
-# Inprint Content 4.5
-# Copyright(c) 2001-2010, Softing, LLC.
+# Inprint Content 5.0
+# Copyright(c) 2001-2011, Softing, LLC.
 # licensing@softing.ru
 # http://softing.ru/license
 
+use utf8;
 use strict;
 use warnings;
 
 use base 'Inprint::BaseController';
+
+sub read {
+    my $c = shift;
+    my $i_id = $c->param("id");
+    
+    my @errors;
+    my $success = $c->json->false;
+
+    push @errors, { id => "id", msg => "Incorrectly filled field"}
+        unless ($c->is_uuid($i_id));
+    
+    my $result = [];
+    
+    unless (@errors) {
+        $result = $c->sql->Q("
+            SELECT id, color, weight as percent, title, shortcut, description, created, updated
+            FROM readiness
+            WHERE id=?
+        ", [ $i_id ])->Hash;
+    }
+    
+    $success = $c->json->true unless (@errors);
+    
+    $c->render_json( { success => $success, errors => \@errors, data => $result } );
+}
 
 sub list {
     my $c = shift;
@@ -30,47 +56,38 @@ sub create {
     my $i_color       = $c->param("color");
     my $i_percent     = $c->param("percent");
 
-    my $result = {
-        success => $c->json->false,
-        errors => []
-    };
+    my @errors;
+    my $success = $c->json->false;
+    
+    push @errors, { id => "title", msg => "Incorrectly filled field"}
+        unless ($c->is_text($i_title));
+        
+    push @errors, { id => "shortcut", msg => "Incorrectly filled field"}
+        unless ($c->is_text($i_shortcut));
+        
+    push @errors, { id => "description", msg => "Incorrectly filled field"}
+        unless ($c->is_text($i_description));
 
-    push @{ $result->{errors} }, { id => "title", msg => "" } unless $i_title;
-    push @{ $result->{errors} }, { id => "shortcut", msg => "" } unless $i_shortcut;
-    push @{ $result->{errors} }, { id => "color", msg => "" } unless $i_color;
-    push @{ $result->{errors} }, { id => "percent", msg => "" } unless $i_percent;
-
-    my $count = 0;
-
-    unless (@{ $result->{errors} }) {
-        $count = $c->sql->Do("
+    push @errors, { id => "color", msg => "Incorrectly filled field"}
+        unless ($c->is_text($i_color));
+        
+    push @errors, { id => "percent", msg => "Incorrectly filled field"}
+        unless ($c->is_int($i_percent));
+    
+    push @errors, { id => "access", msg => "Not enough permissions"}
+        unless ($c->access->Check("domain.readiness.manage"));
+    
+    unless (@errors) {
+        $c->sql->Do("
             INSERT INTO readiness (
                 id, color, weight, title, shortcut, description, created, updated)
             VALUES (?, ?, ?, ?, ?, ?, now(), now());
         ", [ $id, $i_color, $i_percent, $i_title, $i_shortcut, $i_description ]);
     }
-
-    $result->{success} = $c->json->true if $count;
-
-    $c->render_json( $result );
-}
-
-
-sub read {
-    my $c = shift;
-    my $i_id = $c->param("id");
-    my $result = {
-        success => $c->json->false,
-        errors => []
-    };
-    push @{ $result->{errors} }, { id => "id", msg => "" } unless $i_id;
-    $result->{data} = $c->sql->Q("
-        SELECT id, color, weight as percent, title, shortcut, description, created, updated
-        FROM readiness
-        WHERE id=?
-    ", [ $i_id ])->Hash;
-    $result->{success} = $c->json->true if $result->{data};
-    $c->render_json( $result );
+    
+    $success = $c->json->true unless (@errors);
+    
+    $c->render_json({ success => $success, errors => \@errors });
 }
 
 sub update {
@@ -83,41 +100,63 @@ sub update {
     my $i_color       = $c->param("color");
     my $i_percent     = $c->param("percent");
 
-    my $result = {
-        success => $c->json->false,
-        errors => []
-    };
+    my @errors;
+    my $success = $c->json->false;
+    
+    push @errors, { id => "id", msg => "Incorrectly filled field"}
+        unless ($c->is_uuid($i_id));
+    
+    push @errors, { id => "title", msg => "Incorrectly filled field"}
+        unless ($c->is_text($i_title));
+        
+    push @errors, { id => "shortcut", msg => "Incorrectly filled field"}
+        unless ($c->is_text($i_shortcut));
+        
+    push @errors, { id => "description", msg => "Incorrectly filled field"}
+        unless ($c->is_text($i_description));
 
-    push @{ $result->{errors} }, { id => "id", msg => "" } unless $i_id;
-    push @{ $result->{errors} }, { id => "title", msg => "" } unless $i_title;
-    push @{ $result->{errors} }, { id => "shortcut", msg => "" } unless $i_shortcut;
-    push @{ $result->{errors} }, { id => "color", msg => "" } unless $i_color;
-    push @{ $result->{errors} }, { id => "percent", msg => "" } unless $i_percent;
-
-    my $path;
-    my $count = 0;
-
-    unless (@{ $result->{errors} }) {
-        $count = $c->sql->Do("
+    push @errors, { id => "color", msg => "Incorrectly filled field"}
+        unless ($c->is_text($i_color));
+        
+    push @errors, { id => "percent", msg => "Incorrectly filled field"}
+        unless ($c->is_int($i_percent));
+    
+    push @errors, { id => "access", msg => "Not enough permissions"}
+        unless ($c->access->Check("domain.readiness.manage"));
+    
+    unless (@errors) {
+        $c->sql->Do("
             UPDATE readiness SET color=?, weight=?, title=?, shortcut=?, description=?
             WHERE id=? ",
         [ $i_color, $i_percent, $i_title, $i_shortcut, $i_description, $i_id ]);
     }
-
-    $result->{success} = $c->json->true if $count;
-
-    $c->render_json( $result );
+    
+    $success = $c->json->true unless (@errors);
+    
+    $c->render_json({ success => $success, errors => \@errors });
 }
 
 sub delete {
     my $c = shift;
     my @i_ids = $c->param("id");
 
-    foreach my $id (@i_ids) {
-        $c->sql->Do(" DELETE FROM readiness WHERE id =? ", [ $id ]);
+    my @errors;
+    my $success = $c->json->false;
+
+    push @errors, { id => "access", msg => "Not enough permissions"}
+        unless ($c->access->Check("domain.readiness.manage"));
+
+    unless (@errors) {
+        foreach my $id (@i_ids) {
+            if ($c->is_uuid($id)) {
+                $c->sql->Do(" DELETE FROM readiness WHERE id =? ", [ $id ]);
+            }
+        }
     }
 
-    $c->render_json( { } );
+    $success = $c->json->true unless (@errors);
+    
+    $c->render_json({ success => $success, errors => \@errors });
 }
 
 1;
