+++ lib/Inprint/Catalog/Roles.pm
@@ -1,67 +1,93 @@
 package Inprint::Catalog::Roles;
 
-# Inprint Content 4.5
-# Copyright(c) 2001-2010, Softing, LLC.
+# Inprint Content 5.0
+# Copyright(c) 2001-2011, Softing, LLC.
 # licensing@softing.ru
 # http://softing.ru/license
 
+use utf8;
 use strict;
 use warnings;
 
 use base 'Inprint::BaseController';
 
-sub list {
-
+sub read {
     my $c = shift;
+    my $i_id = $c->param("id");
+    
+    my @errors;
+    my $success = $c->json->false;
 
+    push @errors, { id => "id", msg => "Incorrectly filled field"}
+        unless ($c->is_uuid($i_id));
+    
+    my $result = [];
+    
+    unless (@errors) {
+        $result = $c->sql->Q("
+            SELECT id, title, shortcut, description
+            FROM roles
+            WHERE id =?
+        ", [ $i_id ])->Hash;
+    }
+    
+    $success = $c->json->true unless (@errors);
+    
+    $c->render_json( { success => $success, errors => \@errors, data => $result } );
+}
+
+sub list {
+    my $c = shift;
     my $result = $c->sql->Q("
         SELECT id, title, shortcut, description,
             array_to_string(
                 ARRAY(
-                    SELECT rules.shortcut
-                    FROM rules rules, map_role_to_rule mapping
+                    SELECT rules.title
+                    FROM rules, map_role_to_rule mapping
                     WHERE mapping.rule = rules.id AND mapping.role = roles.id
                 )
             , ', ') as rules
         FROM roles
         ORDER BY shortcut
     ")->Hashes;
-
     $c->render_json( { data => $result } );
 }
 
 sub create {
     my $c = shift;
-
+    
     my $id = $c->uuid();
-
+    
     my $i_title       = $c->param("title");
     my $i_shortcut    = $c->param("shortcut");
     my $i_description = $c->param("description");
-
-    $c->sql->Do("
-        INSERT INTO roles(id, title, shortcut, description, created, updated)
-        VALUES (?, ?, ?, ?, now(), now());
-    ", [ $id, $i_title, $i_shortcut, $i_description ]);
-
-    $c->render_json( { success => $c->json->true} );
+    
+    my @errors;
+    my $success = $c->json->false;
+    
+    push @errors, { id => "title", msg => "Incorrectly filled field"}
+        unless ($c->is_text($i_title));
+        
+    push @errors, { id => "shortcut", msg => "Incorrectly filled field"}
+        unless ($c->is_text($i_shortcut));
+        
+    push @errors, { id => "description", msg => "Incorrectly filled field"}
+        unless ($c->is_text($i_description));
+    
+    push @errors, { id => "access", msg => "Not enough permissions"}
+        unless ($c->access->Check("domain.roles.manage"));
+    
+    unless (@errors) {
+        $c->sql->Do("
+            INSERT INTO roles(id, title, shortcut, description, created, updated)
+            VALUES (?, ?, ?, ?, now(), now());
+        ", [ $id, $i_title, $i_shortcut, $i_description ]);
+    }
+    
+    $success = $c->json->true unless (@errors);
+    
+    $c->render_json({ success => $success, errors => \@errors });
 }
-
-sub read {
-
-    my $c = shift;
-
-    my $i_id = $c->param("id");
-
-    my $result = $c->sql->Q("
-        SELECT id, title, shortcut, description
-        FROM roles
-        WHERE id =?
-    ", [ $i_id ])->Hash;
-
-    $c->render_json( { success => $c->json->true, data => $result } );
-}
-
 
 sub update {
     my $c = shift;
@@ -71,22 +97,58 @@ sub update {
     my $i_shortcut    = $c->param("shortcut");
     my $i_description = $c->param("description");
 
-    $c->sql->Do("
-        UPDATE roles
-            SET title=?, shortcut=?, description=?, updated=now()
-        WHERE id =?;
-    ", [ $i_title, $i_shortcut, $i_description, $i_id ]);
+    my @errors;
+    my $success = $c->json->false;
 
-    $c->render_json( { success => $c->json->true} );
+    push @errors, { id => "id", msg => "Incorrectly filled field"}
+        unless ($c->is_uuid($i_id));
+        
+    push @errors, { id => "title", msg => "Incorrectly filled field"}
+        unless ($c->is_text($i_title));
+        
+    push @errors, { id => "shortcut", msg => "Incorrectly filled field"}
+        unless ($c->is_text($i_shortcut));
+        
+    push @errors, { id => "description", msg => "Incorrectly filled field"}
+        unless ($c->is_text($i_description));
+    
+    push @errors, { id => "access", msg => "Not enough permissions"}
+        unless ($c->access->Check("domain.roles.manage"));
+    
+    unless (@errors) {
+        $c->sql->Do("
+            UPDATE roles
+                SET title=?, shortcut=?, description=?, updated=now()
+            WHERE id =?;
+        ", [ $i_title, $i_shortcut, $i_description, $i_id ]);
+    }
+    
+    $success = $c->json->true unless (@errors);
+
+    $c->render_json({ success => $success, errors => \@errors });
 }
 
 sub delete {
     my $c = shift;
     my @ids = $c->param("id");
-    foreach my $id (@ids) {
-        $c->sql->Do(" DELETE FROM roles WHERE id=? ", [ $id ]);
+    
+    my @errors;
+    my $success = $c->json->false;
+    
+    push @errors, { id => "access", msg => "Not enough permissions"}
+        unless ($c->access->Check("domain.roles.manage"));
+    
+    unless (@errors) {
+        foreach my $id (@ids) {
+            if ($c->is_uuid($id)) {
+                $c->sql->Do(" DELETE FROM roles WHERE id=? ", [ $id ]);
+            }
+        }
     }
-    $c->render_json( { success => $c->json->true } );
+    
+    $success = $c->json->true unless (@errors);
+    
+    $c->render_json({ success => $success, errors => \@errors });
 }
 
 
@@ -96,37 +158,48 @@ sub map {
     my $i_id          = $c->param("id");
     my @i_rules       = $c->param("rules");
     my $i_recursive   = $c->param("recursive");
-
-    $c->sql->Do(" DELETE FROM map_role_to_rule WHERE role=? ", [ $i_id ]);
-
-    foreach my $string (@i_rules) {
-        my ($rule, $mode) = split "::", $string;
-        $c->sql->Do("
-            INSERT INTO map_role_to_rule(role, rule, mode)
-                VALUES (?, ?, ?);
-        ", [$i_id, $rule, $mode]);
+    
+    my @errors;
+    my $success = $c->json->false;
+    
+    push @errors, { id => "access", msg => "Not enough permissions"}
+        unless ($c->access->Check("domain.roles.manage"));
+    
+    unless (@errors) {
+        $c->sql->Do(" DELETE FROM map_role_to_rule WHERE role=? ", [ $i_id ]);
+        foreach my $string (@i_rules) {
+            my ($rule, $mode) = split "::", $string;
+            if ($c->is_uuid($i_id) && $c->is_text($rule) && $c->is_text($mode)) {
+                $c->sql->Do(" INSERT INTO map_role_to_rule(role, rule, mode) VALUES (?, ?, ?); ", [$i_id, $rule, $mode]);
+            }
+        }
     }
-
-    $c->render_json( { success => $c->json->true } );
+    
+    $success = $c->json->true unless (@errors);
+    
+    $c->render_json({ success => $success, errors => \@errors });
 }
 
 sub mapping {
 
     my $c = shift;
 
-    my $id = $c->param("id");
+    my $i_id = $c->param("id");
 
-    my $data = $c->sql->Q("
-        SELECT role, rule, mode FROM map_role_to_rule WHERE role =?
-    ", [ $id ])->Hashes;
-
+    my @errors;
+    my $success = $c->json->false;
     my $result = {};
 
-    foreach my $item (@$data) {
-        $result->{ $item->{rule} } = $item->{mode};
+    if ($c->is_uuid($i_id)) {
+        my $data = $c->sql->Q(" SELECT role, rule, mode FROM map_role_to_rule WHERE role =? ", [ $i_id ])->Hashes;
+        foreach my $item (@$data) {
+            $result->{ $item->{rule} } = $item->{mode};
+        }
     }
+    
+    my @ids = $c->param("id");
 
-    $c->render_json( { data => $result } );
+    $c->render_json( { success => $success, errors => \@errors, data => $result } );
 }
 
 1;
